+++ package.json
@@ -76,6 +76,7 @@
     "prettier": "^2.8.8",
     "prop-types": "^15.8.1",
     "react-app-alias": "^2.2.2",
+    "recoil": "^0.7.7",
     "storybook": "^7.0.9",
     "tsconfig-paths-webpack-plugin": "^4.0.1",
     "webpack": "^5.82.0"
